archive/requirements.txt
streamlit
diagrams
openai
python-dotenv
watchdog
